Add port forwarding as its own resource
