+++ openstack/networking/v2/extensions/layer3/floatingips/testing/requests_test.go
@@ -414,59 +414,3 @@ func TestDelete(t *testing.T) {
 	res := floatingips.Delete(fake.ServiceClient(), "2f245a7b-796b-4f26-9cf9-9e82d248fda7")
 	th.AssertNoErr(t, res.Err)
 }
-func TestCreatePortForwarding(t *testing.T) {
-	th.SetupHTTP()
-	defer th.TeardownHTTP()
-
-	th.Mux.HandleFunc("/v2.0/floatingips/2f95fd2b-9f6a-4e8e-9e9a-2cbe286cbf9e/port_forwardings", func(w http.ResponseWriter, r *http.Request) {
-		th.TestMethod(t, r, "POST")
-		th.TestHeader(t, r, "X-Auth-Token", fake.TokenID)
-		th.TestHeader(t, r, "Content-Type", "application/json")
-		th.TestHeader(t, r, "Accept", "application/json")
-		th.TestJSONRequest(t, r, `
-{
-  	"port_forwarding": {
-      "protocol": "tcp",
-      "internal_ip_address": "10.0.0.11",
-      "internal_port": 25,
-      "internal_port_id": "1238be08-a2a8-4b8d-addf-fb5e2250e480",
-      "external_port": 2230
-  }
-}
-		
-			`)
-
-		w.Header().Add("Content-Type", "application/json")
-		w.WriteHeader(http.StatusCreated)
-
-		fmt.Fprintf(w, `
-{
-	"port_forwarding": {
-    		"protocol": "tcp",
-    		"internal_ip_address": "10.0.0.11",
-    		"internal_port": 25,
-    		"internal_port_id": "1238be08-a2a8-4b8d-addf-fb5e2250e480",
-    		"external_port": 2230,
-    		"id": "725ade3c-9760-4880-8080-8fc2dbab9acc"
-  }
-}`)
-	})
-
-	options := floatingips.CreatePortForwardingOpts{
-		Protocol:         "tcp",
-		InternalIPAdress: "10.0.0.11",
-		InternalPort:     25,
-		ExternalPort:     2230,
-		InternalPortID:   "1238be08-a2a8-4b8d-addf-fb5e2250e480",
-	}
-
-	pf, err := floatingips.CreatePortForwarding(fake.ServiceClient(), "2f95fd2b-9f6a-4e8e-9e9a-2cbe286cbf9e", options).ExtractPortForwarding()
-	th.AssertNoErr(t, err)
-
-	th.AssertEquals(t, "725ade3c-9760-4880-8080-8fc2dbab9acc", pf.ID)
-	th.AssertEquals(t, "10.0.0.11", pf.InternalIPAddress)
-	th.AssertEquals(t, 25, pf.InternalPort)
-	th.AssertEquals(t, "1238be08-a2a8-4b8d-addf-fb5e2250e480", pf.InternalPortID)
-	th.AssertEquals(t, 2230, pf.ExternalPort)
-	th.AssertEquals(t, "tcp", pf.Protocol)
-}
