+++ openstack/networking/v2/extensions/layer3/floatingips/urls.go
@@ -3,7 +3,6 @@ package floatingips
 import "github.com/gophercloud/gophercloud"
 
 const resourcePath = "floatingips"
-const portForwardingPath = "port_forwardings"
 
 func rootURL(c *gophercloud.ServiceClient) string {
 	return c.ServiceURL(resourcePath)
@@ -11,12 +10,4 @@ func rootURL(c *gophercloud.ServiceClient) string {
 
 func resourceURL(c *gophercloud.ServiceClient, id string) string {
 	return c.ServiceURL(resourcePath, id)
-}
-
-func portForwardingUrl(c *gophercloud.ServiceClient, id string) string {
-	return c.ServiceURL(resourcePath, id, portForwardingPath)
-}
-
-func singlePortForwardingUrl(c *gophercloud.ServiceClient, id string, portForwardingID string) string {
-	return c.ServiceURL(resourcePath, id, portForwardingPath, portForwardingID)
 }
